EVault/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
Pillow==10.1.0
PyJWT==2.8.0
pymongo==3.12.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.2.4
twilio==8.10.3
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
